requirements.txt
assembla==2.8.0
PyGithub==1.34
